requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.9
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.10.4
python-multipart==0.0.6

Jinja2==3.1.2

numpy==1.25.0
pandas==2.0.3
scikit-learn==1.3.0
transformers==4.30.0

python-dotenv==1.0.0
httpx==0.24.0
pytest==7.4.0
pytest-asyncio==0.21.0

fastapi[all]==0.95.1
